Done: LicenceCheck Based On Admin and Size Check Base On the Licence File

+++ KnowledgeVista/src/main/java/com/knowledgeVista/Course/videoLessons.java
@@ -41,7 +41,7 @@ public class videoLessons {
 	    @Column(nullable = true)
 		private MultipartFile videoFile;
 	    
-	    private Double size;
+	    private Long size;
 	    
 	    @Column(nullable = true)
 	    private String videofilename;
